refactor(sessions): move session types from sessions/ to agent/types
Session type definitions (Session, SessionStore, SessionMetadata, etc.)
were the only content in src/sessions/. Move them into src/agent/types.ts
where the implementation lives, remove the gateway re-exports, and delete
the now-empty sessions/ directory.
