+++ src/test-helpers.ts
@@ -1,7 +1,7 @@
 // src/test-helpers.ts — Shared test mocks for channel tests
 
 import { vi } from "vitest";
-import type { SessionStore } from "./sessions/types.js";
+import type { SessionStore } from "./agent/types.js";
 
 export function createMockSession() {
   let subscriber: ((event: Record<string, unknown>) => void) | null = null;
